+++ changelog.md
@@ -1,4 +1,5 @@
 # PyFit changelog
 **v0.2.0**
-- added automatic rest screen between sets
-- pressing `view workout` won't produce duplicate results anymore
+- Added automatic rest screen between sets
+- Pressing `view workout` won't produce duplicate results anymore
+- Added next exercise info to rest screen