+++ changelog.md
@@ -5,3 +5,4 @@
 - Added next exercise info to rest screen
 - Changed location of default workout to user folder
 - Added the ability to create a new workout file
+- Added the ability to select a stored workout and view it