+++ changelog.md
@@ -1,10 +1,9 @@
 # PyFit changelog
 **v0.2.0**
 - Added automatic rest screen between sets
-- Pressing `view workout` won't produce duplicate results anymore
 - Added next exercise info to rest screen
 - Changed location of default workout to user folder
-- Added the ability to create a new workout file
+- Added the ability to create a new custom workout file
 - Added the ability to select a stored workout and view it
 - Added error message when starting an empty workout
 - Added option to add new steps to a workout